PWMJaguar now extends Jaguar so it can be used in DiscoDrive without issues.
DiscoDrivePWM is now DiscoDriveCAN except for I replaced CAN with PWM everywhere. It hasn't given me any errors, so we should be able to program as we like with PWMJaguar and then later just replace PWM with CAN and everything should work. Though we will probably have to catch some exceptions (PWMJaguar does not throw exceptions). I implemented the soft limits for position control but the hard limit switches still are not implemented. I am not sure if we will use these or not anyway.

DiscoDriveStandard is the old code for normal Jaguars